Updated to V111.1 (login server - done, channel server - only started).
git-svn-id: https://xp-dev.com/svn/MapleGame@59 22f4eea3-df18-4490-a9d7-79e6856d3c20

+++ ChannelServer/Maple/Life/Mob.cs
@@ -46,7 +46,7 @@ namespace Loki.Maple.Life
         public int DropItemPeriod { get; private set; }
         public byte HpBarForeColor { get; private set; }
         public byte HpBarBackColor { get; private set; }
-        public byte CarnivalPoints { get; private set; }
+        public ushort CarnivalPoints { get; private set; }
         public int WeaponAttack { get; private set; }
         public int WeaponDefense { get; private set; }
         public int MagicAttack { get; private set; }
